pyproject.toml
[tool.poetry]
name = "bulletinboard"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "5.2.11"
djangorestframework = ">=3.16.1,<4.0.0"
celery = ">=5.6.2,<6.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
djangorestframework-simplejwt = ">=5.5.1,<6.0.0"
django-filter = ">=25.2,<26.0"
django-cors-headers = ">=4.9.0,<5.0.0"
django-celery-beat = "2.1.0"
psycopg2-binary = ">=2.9.11,<3.0.0"
requests = ">=2.32.5,<3.0.0"
pillow = ">=12.1.1,<13.0.0"
redis = "7"
gevent = "^25.9.1"
pytest = "^9.0.2"
drf-yasg = "^1.21.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)"
]

[tool.black]
line-length = 119
exclude = ".git"

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.pytest.ini_options]
pythonpath = [
  "."
]